+++ ajc27_freemocap_blender_addon/core_functions/create_video/helpers/overlay_components/frame_number_overlay.py
@@ -1,12 +1,13 @@
 import numpy as np
-from freemocap_video_export.create_video.visual_overlays.frame_information_dataclass import FrameInformation
-from freemocap_video_export.video_config import visual_components
+
+from ajc27_freemocap_blender_addon.core_functions.create_video.helpers.overlay_components.frame_information_dataclass import FrameInformation
+from ajc27_freemocap_blender_addon.data_models.parameter_models.video_config import VISUAL_COMPONENTS
 
 
-class VisualComponentFrameNumber:
-    def __init__(self, **kwargs):
-        self.position_x_pct = visual_components['frame_number']['position_x_pct']
-        self.position_y_pct = visual_components['frame_number']['position_y_pct']
+class OverlayFrameNumber:
+    def __init__(self, frame_info=None):
+        self.position_x_pct = VISUAL_COMPONENTS['frame_number']['position_x_pct']
+        self.position_y_pct = VISUAL_COMPONENTS['frame_number']['position_y_pct']
 
     def add_component(self,
                       image: np.ndarray,
@@ -17,11 +18,11 @@ class VisualComponentFrameNumber:
             image,
             str(frame_info.frame_number).zfill(frame_info.total_frames_digits) + '/' + str(frame_info.total_frames),
             (int(self.position_x_pct * frame_info.width), int(self.position_y_pct * frame_info.height)),
-            visual_components['frame_number']['font'],
-            visual_components['frame_number']['fontScale'],
-            visual_components['frame_number']['color'],
-            visual_components['frame_number']['thickness'],
-            visual_components['frame_number']['lineType'],
+            VISUAL_COMPONENTS['frame_number']['font'],
+            VISUAL_COMPONENTS['frame_number']['fontScale'],
+            VISUAL_COMPONENTS['frame_number']['color'],
+            VISUAL_COMPONENTS['frame_number']['thickness'],
+            VISUAL_COMPONENTS['frame_number']['lineType'],
             )
 
         return annotated_image
