+++ ajc27_freemocap_blender_addon/core_functions/create_video/helpers/overlay_components/matplotlib_plot_overlays.py
@@ -1,10 +1,12 @@
+import os
 import bpy
 import numpy as np
-from freemocap_video_export.create_video.visual_overlays.frame_information_dataclass import FrameInformation
-from freemocap_video_export.video_config import visual_components, color_palette
+
+from ajc27_freemocap_blender_addon.core_functions.create_video.helpers.overlay_components.frame_information_dataclass import FrameInformation
+from ajc27_freemocap_blender_addon.data_models.parameter_models.video_config import VISUAL_COMPONENTS, COLOR_PALETTE
 
 
-class VisualComponentPlotComBos:
+class OverlayPlotComBos:
     def __init__(self, frame_info: FrameInformation):
         self.scene = frame_info.scene
 
@@ -53,45 +55,44 @@ class VisualComponentPlotComBos:
         ]
 
         ### Plot setup ###
-
         # Create the plot with size according to the component size
         fig, ax = plt.subplots(figsize=(6.4,
-                                        6.4 * frame_info.height * visual_components['VisualComponentComBos']['height_pct'] / (
-                                                frame_info.width * visual_components['VisualComponentComBos']['width_pct'])))
+                                        6.4 * frame_info.height * VISUAL_COMPONENTS['plot_com_bos']['height_pct'] / (
+                                                frame_info.width * VISUAL_COMPONENTS['plot_com_bos']['width_pct'])))
 
         # Plot the COM
-        ax.scatter(0, 0, marker='o', color=color_palette['dark_terra_cotta']['hex'], zorder=2)
+        ax.scatter(0, 0, marker='o', color=COLOR_PALETTE['dark_terra_cotta']['hex'], zorder=2)
 
         # Filter out points that are above the ground contact threshold
         filtered_base_of_support_points = {
             'x': [x for x, z in zip(base_of_support_points['x'], base_of_support_points['z']) if
-                  z < visual_components['VisualComponentComBos']['ground_contact_threshold']],
+                  z < VISUAL_COMPONENTS['plot_com_bos']['ground_contact_threshold']],
             'y': [y for y, z in zip(base_of_support_points['y'], base_of_support_points['z']) if
-                  z < visual_components['VisualComponentComBos']['ground_contact_threshold']],
+                  z < VISUAL_COMPONENTS['plot_com_bos']['ground_contact_threshold']],
             'z': [z for z in base_of_support_points['z'] if
-                  z < visual_components['VisualComponentComBos']['ground_contact_threshold']],
+                  z < VISUAL_COMPONENTS['plot_com_bos']['ground_contact_threshold']],
         }
 
         # Plot the base of support
         ax.scatter(filtered_base_of_support_points['x'],
                    filtered_base_of_support_points['y'],
                    marker='o',
-                   color=color_palette['crystal']['hex'],
+                   color=COLOR_PALETTE['crystal']['hex'],
                    zorder=2,
                    )
 
         # Add labels
-        plt.text(0, 0, 'COM', color=color_palette['crystal']['hex'], fontsize=12)
+        plt.text(0, 0, 'COM', color=COLOR_PALETTE['crystal']['hex'], fontsize=12)
 
         # Filter the labels based on the filtered base of support points
         filtered_scatter_labels = [label for label, z in zip(scatter_labels, base_of_support_points['z']) if
-                                   z < visual_components['VisualComponentComBos']['ground_contact_threshold']]
+                                   z < VISUAL_COMPONENTS['plot_com_bos']['ground_contact_threshold']]
 
         for point in range(0, len(filtered_base_of_support_points['x'])):
             plt.text(filtered_base_of_support_points['x'][point],
                      filtered_base_of_support_points['y'][point],
                      filtered_scatter_labels[point],
-                     color=color_palette['crystal']['hex'],
+                     color=COLOR_PALETTE['crystal']['hex'],
                      fontsize=12)
 
         # Connect the points to form a polygon if there are more than 1 point
@@ -115,10 +116,10 @@ class VisualComponentPlotComBos:
         ax.set_ylim(-max_value * 1.1, max_value * 1.1)
 
         # Set the title and axes labels
-        ax.set_title('COM v/s BOS', color=color_palette['crystal']['hex'], fontsize=30)
-        ax.set_xlabel('X Position', color=color_palette['crystal']['hex'], fontsize=24)
-        ax.set_ylabel('Y Position', color=color_palette['crystal']['hex'], fontsize=24)
-        ax.tick_params(axis='both', which='both', labelsize=14, colors=color_palette['dark_terra_cotta']['hex'])
+        ax.set_title('COM v/s BOS', color=COLOR_PALETTE['crystal']['hex'], fontsize=30)
+        ax.set_xlabel('X Position', color=COLOR_PALETTE['crystal']['hex'], fontsize=24)
+        ax.set_ylabel('Y Position', color=COLOR_PALETTE['crystal']['hex'], fontsize=24)
+        ax.tick_params(axis='both', which='both', labelsize=14, colors=COLOR_PALETTE['dark_terra_cotta']['hex'])
 
         # Enable grid lines
         plt.grid(True)
@@ -127,36 +128,38 @@ class VisualComponentPlotComBos:
         ax.grid(color='gray', linestyle='--', linewidth=0.5)
 
         # Set the axes colors
-        ax.spines['bottom'].set_color(color_palette['crystal']['hex'])
-        ax.spines['top'].set_color(color_palette['crystal']['hex'])
-        ax.spines['left'].set_color(color_palette['crystal']['hex'])
-        ax.spines['right'].set_color(color_palette['crystal']['hex'])
+        ax.spines['bottom'].set_color(COLOR_PALETTE['crystal']['hex'])
+        ax.spines['top'].set_color(COLOR_PALETTE['crystal']['hex'])
+        ax.spines['left'].set_color(COLOR_PALETTE['crystal']['hex'])
+        ax.spines['right'].set_color(COLOR_PALETTE['crystal']['hex'])
 
         # Adjust the padding around the plot area
         plt.subplots_adjust(left=0.18, right=0.95, bottom=0.2, top=0.85)
 
         # Set the background color and transparency
-        ax.set_facecolor(color_palette['japanese_indigo']['hex'])
+        ax.set_facecolor(COLOR_PALETTE['japanese_indigo']['hex'])
         ax.patch.set_alpha(0.7)
-        fig.set_facecolor(color_palette['japanese_indigo']['hex'])
+        fig.set_facecolor(COLOR_PALETTE['japanese_indigo']['hex'])
         fig.patch.set_alpha(0.7)
 
         ### Plot image export ###
 
         # Convert the Matplotlib plot to an OpenCV image.
         fig.canvas.draw()
+        # Create the directory if it doesn't exist
+        os.makedirs(os.path.dirname(str(frame_info.file_directory) + '/video/'), exist_ok=True)
+        # Save the image
         fig.savefig(str(frame_info.file_directory) + '/video/' + 'plot_aux.png', transparent=False)
 
         # Close the figure
         plt.close(fig)
 
         ### Image appending ###
-
         # Read the image component
         img = cv2.imread(str(frame_info.file_directory) + '/video/' + 'plot_aux.png', cv2.IMREAD_UNCHANGED)
 
         # Resize the component
-        component = visual_components['VisualComponentComBos']
+        component = VISUAL_COMPONENTS['plot_com_bos']
         img_resized = cv2.resize(img,
                                  (int(frame_info.width * component['width_pct']),
                                   int(frame_info.height * component['height_pct'])),
@@ -185,7 +188,7 @@ class VisualComponentPlotComBos:
         return image
 
 
-class VisualComponentPlotFootDeviation:
+class OverlayPlotFootDeviation:
     def __init__(self, frame_info: FrameInformation):
         self.scene = frame_info.scene
 
@@ -193,7 +196,8 @@ class VisualComponentPlotFootDeviation:
                       image: np.ndarray,
                       frame_info: FrameInformation):
         ### Data setup ###
-
+        from matplotlib import pyplot as plt
+        import cv2
         # Set the frame according to the frame number
         self.scene.frame_set(frame_info.frame_start + frame_info.frame_number)
 
@@ -241,47 +245,46 @@ class VisualComponentPlotFootDeviation:
         left_foot_index_y = foot_length * np.cos(left_hip_foot_angle)
 
         ### Plot setup ###
-
         # Create the plot objects
-        fig, ax = plt.subplots(figsize=(6.4, 6.4 * frame_info.height * visual_components['VisualComponentFootDeviation'][
-            'height_pct'] / (frame_info.width * visual_components['VisualComponentFootDeviation']['width_pct'])))
+        fig, ax = plt.subplots(figsize=(6.4, 6.4 * frame_info.height * VISUAL_COMPONENTS['plot_foot_deviation'][
+            'height_pct'] / (frame_info.width * VISUAL_COMPONENTS['plot_foot_deviation']['width_pct'])))
 
         # Plot the fixed points
         for fixed_point in fixed_points:
             ax.scatter(fixed_points[fixed_point][0], fixed_points[fixed_point][1], marker='o',
-                       color=color_palette['dark_terra_cotta']['hex'], zorder=2)
+                       color=COLOR_PALETTE['dark_terra_cotta']['hex'], zorder=2)
 
         # Add labels
         plt.text(fixed_points['right_foot_origin'][0], fixed_points['right_foot_origin'][1] - 12, 'RIGHT',
-                 color=color_palette['crystal']['hex'], fontsize=23, ha='center')
+                 color=COLOR_PALETTE['crystal']['hex'], fontsize=23, ha='center')
         plt.text(fixed_points['left_foot_origin'][0], fixed_points['left_foot_origin'][1] - 12, 'LEFT',
-                 color=color_palette['crystal']['hex'], fontsize=23, ha='center')
+                 color=COLOR_PALETTE['crystal']['hex'], fontsize=23, ha='center')
         plt.text(fixed_points['right_foot_origin'][0], fixed_points['right_foot_origin'][1] - 5,
-                 str(np.around(np.degrees(right_hip_foot_angle), 1)) + '°', color=color_palette['crystal']['hex'],
+                 str(np.around(np.degrees(right_hip_foot_angle), 1)) + '°', color=COLOR_PALETTE['crystal']['hex'],
                  fontsize=22, ha='center')
         plt.text(fixed_points['left_foot_origin'][0], fixed_points['left_foot_origin'][1] - 5,
-                 str(np.around(np.degrees(left_hip_foot_angle), 1)) + '°', color=color_palette['crystal']['hex'],
+                 str(np.around(np.degrees(left_hip_foot_angle), 1)) + '°', color=COLOR_PALETTE['crystal']['hex'],
                  fontsize=22, ha='center')
         plt.text(fixed_points['right_foot_origin'][0], fixed_points['right_foot_origin'][1] + 35, '0',
-                 color=color_palette['crystal']['hex'], fontsize=12, ha='center')
+                 color=COLOR_PALETTE['crystal']['hex'], fontsize=12, ha='center')
         plt.text(fixed_points['left_foot_origin'][0], fixed_points['left_foot_origin'][1] + 35, '0',
-                 color=color_palette['crystal']['hex'], fontsize=12, ha='center')
+                 color=COLOR_PALETTE['crystal']['hex'], fontsize=12, ha='center')
         plt.text(fixed_points['right_foot_origin'][0] - 35, fixed_points['right_foot_origin'][1], '90',
-                 color=color_palette['crystal']['hex'], fontsize=12, ha='right', va='center')
+                 color=COLOR_PALETTE['crystal']['hex'], fontsize=12, ha='right', va='center')
         plt.text(fixed_points['left_foot_origin'][0] + 35, fixed_points['left_foot_origin'][1], '90',
-                 color=color_palette['crystal']['hex'], fontsize=12, ha='left', va='center')
-        plt.text(0, 0, '-90', color=color_palette['crystal']['hex'], fontsize=12, ha='center', va='center')
+                 color=COLOR_PALETTE['crystal']['hex'], fontsize=12, ha='left', va='center')
+        plt.text(0, 0, '-90', color=COLOR_PALETTE['crystal']['hex'], fontsize=12, ha='center', va='center')
 
         # Plot the line from foot origin to the foot index
         plt.plot([fixed_points['right_foot_origin'][0], fixed_points['right_foot_origin'][0] + right_foot_index_x],
                  [fixed_points['right_foot_origin'][1], fixed_points['right_foot_origin'][1] + right_foot_index_y],
-                 color=color_palette['dark_terra_cotta']['hex'],
+                 color=COLOR_PALETTE['dark_terra_cotta']['hex'],
                  linewidth=5,
                  zorder=1,
                  )
         plt.plot([fixed_points['left_foot_origin'][0], fixed_points['left_foot_origin'][0] + left_foot_index_x],
                  [fixed_points['left_foot_origin'][1], fixed_points['left_foot_origin'][1] + left_foot_index_y],
-                 color=color_palette['dark_terra_cotta']['hex'],
+                 color=COLOR_PALETTE['dark_terra_cotta']['hex'],
                  linewidth=5,
                  zorder=1,
                  )
@@ -290,22 +293,21 @@ class VisualComponentPlotFootDeviation:
         ax.scatter(fixed_points['right_foot_origin'][0] + right_foot_index_x,
                    fixed_points['right_foot_origin'][1] + right_foot_index_y,
                    marker='o',
-                   color=color_palette['crystal']['hex'],
+                   color=COLOR_PALETTE['crystal']['hex'],
                    zorder=3,
                    s=50,
                    )
         ax.scatter(fixed_points['left_foot_origin'][0] + left_foot_index_x,
                    fixed_points['left_foot_origin'][1] + left_foot_index_y,
                    marker='o',
-                   color=color_palette['crystal']['hex'],
+                   color=COLOR_PALETTE['crystal']['hex'],
                    zorder=3,
                    s=50,
                    )
 
         ### Plot format setup ###
-
         # Set plot title
-        ax.set_title('Foot Deviation', color=color_palette['crystal']['hex'], fontsize=30)
+        ax.set_title('Foot Deviation', color=COLOR_PALETTE['crystal']['hex'], fontsize=30)
 
         # Set the axes limits
         ax.set_xlim(-(foot_length * 2 + x_axis_internal_margin + x_axis_external_margin),
@@ -323,31 +325,32 @@ class VisualComponentPlotFootDeviation:
         plt.gca().set_frame_on(False)
 
         # Set the axes colors
-        ax.spines['bottom'].set_color(color_palette['crystal']['hex'])
-        ax.spines['top'].set_color(color_palette['crystal']['hex'])
-        ax.spines['left'].set_color(color_palette['crystal']['hex'])
-        ax.spines['right'].set_color(color_palette['crystal']['hex'])
+        ax.spines['bottom'].set_color(COLOR_PALETTE['crystal']['hex'])
+        ax.spines['top'].set_color(COLOR_PALETTE['crystal']['hex'])
+        ax.spines['left'].set_color(COLOR_PALETTE['crystal']['hex'])
+        ax.spines['right'].set_color(COLOR_PALETTE['crystal']['hex'])
 
         # Set the background color and transparency
-        fig.set_facecolor(color_palette['japanese_indigo']['hex'])
+        fig.set_facecolor(COLOR_PALETTE['japanese_indigo']['hex'])
         fig.patch.set_alpha(0.7)
 
         ### Plot image export ###
-
         # Convert the Matplotlib plot to an OpenCV image.
         fig.canvas.draw()
+        # Create the directory if it doesn't exist
+        os.makedirs(os.path.dirname(str(frame_info.file_directory) + '/video/'), exist_ok=True)
+        # Save the image
         fig.savefig(str(frame_info.file_directory) + '/video/' + 'plot_aux.png', transparent=False, bbox_inches='tight')
 
         # Close the figure
         plt.close(fig)
 
         ### Image appending ###
-
         # Read the image component
         img = cv2.imread(str(frame_info.file_directory) + '/video/' + 'plot_aux.png', cv2.IMREAD_UNCHANGED)
 
         # Resize the component
-        component = visual_components['VisualComponentFootDeviation']
+        component = VISUAL_COMPONENTS['plot_foot_deviation']
         img_resized = cv2.resize(img,
                                  (int(frame_info.width * component['width_pct']),
                                   int(frame_info.height * component['height_pct'])),
