+++ ajc27_freemocap_blender_addon/core_functions/create_video/helpers/rearrange_background_videos.py
@@ -1,9 +1,10 @@
 import bpy
 
 def rearrange_background_videos(
-        scene: bpy.types.Scene = None,
-        videos_x_separation: float = 0.1
+    scene: bpy.types.Scene=None,
+    videos_x_separation: float=0.1
 ) -> None:
+
     # Create a list with the background videos
     background_videos = []
 
@@ -16,10 +17,18 @@ def rearrange_background_videos(
     videos_x_dimension = background_videos[0].dimensions.x
 
     # Calculate the first video x position (from the left to the right)
-    first_video_x_position = -(len(background_videos) - 1) / 2 * (videos_x_dimension + videos_x_separation)
+    first_video_x_position = (
+        -(len(background_videos) - 1)
+        / 2 * (videos_x_dimension + videos_x_separation)
+    )
 
     # Iterate through the background videos
     for video_index in range(len(background_videos)):
         # Set the location of the video
-        background_videos[video_index].location[0] = first_video_x_position + video_index * (
-                    videos_x_dimension + videos_x_separation)
+        background_videos[video_index].location[0] = (
+            first_video_x_position
+            + video_index
+            * (videos_x_dimension + videos_x_separation)
+        )
+
+    return
