+++ ajc27_freemocap_blender_addon/core_functions/create_video/helpers/reset_scene_defaults.py
@@ -1,0 +1,15 @@
+import bpy
+
+def reset_scene_defaults() -> None:
+
+    # Enable all elements in render
+    for obj in bpy.data.objects:
+        obj.hide_render = False
+
+    # Hide the background if present
+    if "background" in bpy.data.objects:
+        bpy.data.objects["background"].hide_set(True)
+    else:
+        print("The 'background' object is not present in the scene.")
+
+    return
